Title: 합과 차의 나눗셈, Time: 103ms, Memory: 24MB, Status: Wrong Answer - CodeTree

+++ 240802/합과 차의 나눗셈/divide-of-sum-and-sub.py
@@ -1,4 +1,4 @@
 a,b= map(int,input().split())
 c = a+b
 d = a-b
-print(round(float(c/d),2))
+print(round((c/d),2))
